antes que se me vaya la luz denuevo

+++ AHM_TOTAL_TRAVEL_APP/flutter_application_1/lib/Models/ReservationViewModel.dart
@@ -107,7 +107,7 @@ class ReservHotelModel {
   String? reHoFechaSalida;
   int? resvId;
   int? hoteId;
-  int? reHoPrecioTotal;
+  double? reHoPrecioTotal;
   int? reHoUsuarioCreacion;
   int? reHoUsuarioModifica;
